+++ srcs/builtins/type.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   type.c                                             :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: mpivet-p <mpivet-p@student.42.fr>          +#+  +:+       +#+        */
+/*   By: arsciand <arsciand@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/12/16 05:41:09 by mpivet-p          #+#    #+#             */
-/*   Updated: 2020/02/16 22:42:39 by mpivet-p         ###   ########.fr       */
+/*   Updated: 2020/03/08 21:19:37 by arsciand         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -41,8 +41,8 @@ static int8_t	type(t_core *shell, char *path)
 
 int8_t			builtin_type(t_core *shell, t_process *process)
 {
-	int		i;
-	int		ret;
+	size_t	i;
+	int8_t	ret;
 
 	i = 1;
 	ret = 0;
