+++ srcs/exec/call_builtin.c
@@ -6,7 +6,7 @@
 /*   By: arsciand <arsciand@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/09/30 00:24:24 by mpivet-p          #+#    #+#             */
-/*   Updated: 2020/03/08 17:27:08 by arsciand         ###   ########.fr       */
+/*   Updated: 2020/03/08 21:35:02 by arsciand         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -28,7 +28,7 @@ int8_t	is_a_blt(const char *cmd)
 	return (FAILURE);
 }
 
-int8_t	call_builtin(t_core *shell, t_process *process, int blt)
+int8_t	call_builtin(t_core *shell, t_process *process, int8_t blt)
 {
 	static int8_t	(*blt_call[14])(t_core *, t_process *) = {
 		builtin_set, builtin_unset, builtin_export, builtin_exit, builtin_cd
